requirements.txt
pandas>=1.5.0
openreview-py>=1.0.0
tqdm>=4.64.0
pyarrow>=10.0.0
requests>=2.28.0
spacy>=3.4.0

